Tested generate new hands. Passing tests
